refactor: projectCrews contents 분리
