--- src/components/ui/skeleton/project/crews/CrewListSkeleton.tsx
@@ -1,15 +1,0 @@
-import Skeleton from '@/components/ui/skeleton/Skeleton';
-
-const CrewListSkeleton = () => {
-  return (
-    <ul role='list' className='w-full space-y-2'>
-      <Skeleton className='h-[70px]' />
-      <Skeleton className='h-[70px]' />
-      <Skeleton className='h-[70px]' />
-      <Skeleton className='h-[70px]' />
-      <Skeleton className='h-[70px]' />
-    </ul>
-  );
-};
-
-export default CrewListSkeleton;
