--- src/features/project/auth/projectCrews/components/Crews.tsx
@@ -1,19 +1,0 @@
-'use client';
-
-import CrewList from '@/features/project/auth/projectCrews/components/CrewList';
-import CrewListSkeleton from '@/components/ui/skeleton/project/crews/CrewListSkeleton';
-import { Suspense } from 'react';
-
-const Crews = () => {
-  return (
-    <section className='w-full flex flex-col items-center px-1'>
-      <section className='w-full mobile:max-h-[400px] mx-auto  mobile:overflow-y-scroll'>
-        <Suspense fallback={<CrewListSkeleton />}>
-          <CrewList />
-        </Suspense>
-      </section>
-    </section>
-  );
-};
-
-export default Crews;
