+++ src/features/project/auth/projectCrews/store/CrewFWModalStateStore.tsx
@@ -1,25 +1,21 @@
 import { ModalState } from '@/utils/type';
 import { atom } from 'recoil';
-import { ProjectAuthCode } from '@/features/project/auth/projectManageAuth/types/projectAuth';
-import { PROJECT_AUTH_CODE } from '@/features/project/auth/projectManageAuth/constants/projectAuthCode';
 import { CreateCrewFWVoteReason } from '@/features/project/auth/projectVote/service/createFWVote';
 
 export type CrewFWModalState = ModalState & {
-  projectId: bigint;
-  crewId: bigint;
-  crewPMAuth: ProjectAuthCode;
-  userPMAuth: ProjectAuthCode;
+  projectId: string;
+  crewId: string;
+  crewPMAuth: string;
+  userPMAuth: string;
 };
-
-const { CREW: CREW_AUTH } = PROJECT_AUTH_CODE;
 
 export const DEFAULT_FW_MODAL_STATE: CrewFWModalState = {
   isOpen: false,
   title: '강제탈퇴 투표 신청',
-  projectId: 0n,
-  crewId: 0n,
-  crewPMAuth: CREW_AUTH,
-  userPMAuth: CREW_AUTH,
+  projectId: '0',
+  crewId: '0',
+  crewPMAuth: '',
+  userPMAuth: '',
 };
 
 export const crewFWModalStateStore = atom<CrewFWModalState>({
